// src/types/supplier.ts
export default interface Supplier {
  id: string;
  name: string;
  description: string;
  logo_image: string;
  marketing_image: string;
  social_facebook: string;
  social_instagram: string;
  social_linkedin: string;
  social_twitter: string;
  social_youtube: string;
  social_website: string;
}